+++ .gitignore
@@ -1,3 +1,5 @@
 /target
 
+.vscode/
+
 .env
