+++ src/resources/devices/devices.rs
@@ -2,3 +2,4 @@ pub mod battery;
 pub mod meter;
 pub mod solar;
 pub mod thermal;
+pub mod ha_entity;
