+++ src/resources/house.rs
@@ -1,8 +1,8 @@
-use actix_web::{get, post, web, HttpResponse, Responder};
+use actix_web::{get, web, HttpResponse, Responder};
 
 #[path = "devices/devices.rs"]
 mod devices;
-use devices::{battery, meter, solar, thermal};
+use devices::{battery, ha_entity, meter, solar, thermal};
 
 pub fn configure(cfg: &mut web::ServiceConfig) {
     cfg.service(
@@ -11,7 +11,8 @@ pub fn configure(cfg: &mut web::ServiceConfig) {
             .configure(battery::configure)
             .configure(meter::configure)
             .configure(solar::configure)
-            .configure(thermal::configure),
+            .configure(thermal::configure)
+            .configure(ha_entity::configure),
     );
 }
